Move Dropout and Multinomial shape inference into a random submodule
Move the Dropout and Multinomial operators, plus their tests, from ops.rs
into a new ops/random.rs submodule, matching how these operators are
grouped in src/ops.rs.
